First upload
Draggable class has a JComponent that can be moved around the screen with right click and plays a sound on left click. The recorder class records input from a microphone and saves it to file...Unsure as to how it works but video from where i got it is linked. It seems to use complex bitrate and audio recording things which are simplified with existing java classes.